Move DigitalTwinTests to right folder. (#1122)
#1017
